update model and service

+++ model/request/file.go
@@ -3,7 +3,7 @@ package request
 type File struct {
 	Name     string `json:"name" binding:"required"`
 	ParentID string `json:"parent_id" binding:"required"`
-	Size     int    `json:"size" binding:"required"`
+	Size     uint64 `json:"size" binding:"required"`
 	SHA256   string `json:"sha256" binding:"required"`
 	Provider string `json:"provider" binding:"required"`
 }
@@ -11,7 +11,7 @@ type File struct {
 type PreFile struct {
 	Name     string `json:"name" binding:"required"`
 	ParentID string `json:"parent_id" binding:"required"`
-	Size     int    `json:"size" binding:"required"`
+	Size     uint64 `json:"size" binding:"required"`
 	SHA256   string `json:"sha256" binding:"required"`
 }
